WasmFX httpserver implementation working with all optimisation levels
This commit makes the WasmFX powered httpserver work with all
optimisation levels of clang. The problem is that stack switching
violates the ABI of Wasm generated by clang. In particular, the
generated code maintains a shadow stack which gets out of sync with
the actual Wasm stack when we stack switch. To code around this issue,
we lift (most) state to the global scope.